+++ src/module/Entity/config/module.config.php
@@ -32,6 +32,7 @@ return [
     'class_resolver' => [
         'Entity\Entity\EntityInterface' => 'Entity\Entity\Entity',
         'Entity\Entity\TypeInterface' => 'Entity\Entity\Type',
+        'Entity\Entity\RevisionInterface' => 'Entity\Entity\Revision',
     ],
     'di' => [
         'allowed_controllers' => [