+++ src/module/Subject/src/Subject/Factory/SubjectManagerFactory.php
@@ -28,8 +28,9 @@ class SubjectManagerFactory implements FactoryInterface
     {
         $taxonomyManager = $this->getTaxonomyManager($serviceLocator);
         $storage         = $serviceLocator->get('Subject\Storage\SubjectStorage');
-        $normalizer = $serviceLocator->get('Normalizer\Normalizer');
-        $service         = new SubjectManager($normalizer, $storage, $taxonomyManager);
+        $normalizer      = $serviceLocator->get('Normalizer\Normalizer');
+        $entityManager   = $serviceLocator->get('Entity\Manager\EntityManager');
+        $service         = new SubjectManager($normalizer, $storage, $taxonomyManager, $entityManager);
 
         return $service;
     }