+++ src/module/Subject/src/Subject/Manager/SubjectManager.php
@@ -12,41 +12,50 @@ namespace Subject\Manager;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Entity\Entity\EntityInterface;
+use Entity\Entity\RevisionInterface;
 use Instance\Entity\InstanceInterface;
 use Normalizer\Normalizer;
 use Taxonomy\Entity\TaxonomyTermInterface;
 use Taxonomy\Manager\TaxonomyManagerAwareTrait;
 use Taxonomy\Manager\TaxonomyManagerInterface;
+use Versioning\Entity\RepositoryInterface;
 use Zend\Cache\Storage\StorageInterface;
+use Entity\Manager\EntityManagerInterface;
 
 class SubjectManager implements SubjectManagerInterface
 {
     use TaxonomyManagerAwareTrait;
 
     /**
+     *
      * @var StorageInterface
      */
     protected $storage;
 
     /**
+     *
      * @var Normalizer
      */
     protected $normalizer;
 
-    public function __construct(
-        Normalizer $normalizer,
-        StorageInterface $storage,
-        TaxonomyManagerInterface $taxonomyManager
-    ) {
+    /**
+     *
+     * @var EntityManagerInterface
+     */
+    protected $entityManager;
+
+    public function __construct(Normalizer $normalizer, StorageInterface $storage, TaxonomyManagerInterface $taxonomyManager, EntityManagerInterface $entityManager)
+    {
         $this->taxonomyManager = $taxonomyManager;
-        $this->storage         = $storage;
-        $this->normalizer      = $normalizer;
+        $this->storage = $storage;
+        $this->normalizer = $normalizer;
+        $this->entityManager = $entityManager;
     }
 
     public function findSubjectByString($name, InstanceInterface $instance)
     {
         $taxonomy = $this->getTaxonomyManager()->findTaxonomyByName('subject', $instance);
-        $term     = $this->getTaxonomyManager()->findTermByName($taxonomy, (array)$name);
+        $term = $this->getTaxonomyManager()->findTermByName($taxonomy, (array) $name);
         return $term;
     }
 
@@ -64,46 +73,16 @@ class SubjectManager implements SubjectManagerInterface
 
     public function getTrashedEntities(TaxonomyTermInterface $term)
     {
-        $key = 'trashed:' .hash('sha256', serialize($term));
+        $key = 'trashed:' . hash('sha256', serialize($term));
         if ($this->storage->hasItem($key)) {
             return $this->storage->getItem($key);
         }
 
-        $entities   = $this->getEntities($term);
+        $entities = $this->getEntities($term);
         $collection = new ArrayCollection();
         $this->iterEntities($entities, $collection, 'isTrashed');
         $this->storage->setItem($key, $collection);
         return $collection;
-    }
-
-    public function getUnrevisedRevisions(TaxonomyTermInterface $term)
-    {
-        $key = 'unrevised:' . hash('sha256', serialize($term));
-        if ($this->storage->hasItem($key)) {
-            return $this->storage->getItem($key);
-        }
-
-        $entities   = $this->getEntities($term);
-        $collection = new ArrayCollection();
-        $this->iterEntities($entities, $collection, 'isRevised');
-        $iterator = $collection->getIterator();
-        $iterator->ksort();
-        $collection = new ArrayCollection(iterator_to_array($iterator));
-        $this->storage->setItem($key, $collection);
-        return $collection;
-    }
-
-    protected function getEntities(TaxonomyTermInterface $term)
-    {
-        return $term->getAssociatedRecursive('entities');
-    }
-
-    protected function isRevised(EntityInterface $entity, Collection $collection)
-    {
-        if ($entity->isUnrevised() && !$collection->contains($entity)) {
-            $normalized = $this->normalizer->normalize($entity->getHead());
-            $collection->set(-$normalized->getMetadata()->getCreationDate()->getTimestamp(), $normalized);
-        }
     }
 
     protected function isTrashed(EntityInterface $entity, Collection $collection)
@@ -113,6 +92,69 @@ class SubjectManager implements SubjectManagerInterface
             $normalized = $this->normalizer->normalize($entity);
             $collection->add($normalized);
         }
+    }
+
+    public function getUnrevisedRevisions(TaxonomyTermInterface $term)
+    {
+        $revisions = $this->entityManager->findAllUnrevisedRevisions();
+
+        // collection for filtered entities by $term
+        $filteredRevisions = new ArrayCollection();
+        $recentTimestampsPerEntity = new ArrayCollection();
+
+        // find all entities where $term matches (also in parents)
+        foreach ($revisions as $revision) {
+            if ($this->isInSubject($revision->getRepository(), $term)) {
+                $normalized = $this->normalizer->normalize($revision);
+                $filteredRevisions->add($revision);
+                $entityId = $revision->getRepository()->getId();
+                $recentTimestampsPerEntity->set(
+                    $entityId,
+                    max($normalized->getMetadata()->getCreationDate()->getTimestamp(), $recentTimestampsPerEntity->get($entityId))
+                );
+            }
+        }
+
+        $iterator = $filteredRevisions->getIterator();
+        $iterator->uasort(function ($revisionA, $revisionB) use ($recentTimestampsPerEntity) {
+            /**
+             * @var RevisionInterface $revisionA
+             * @var RevisionInterface $revisionB
+             */
+            $entityA = $revisionA->getRepository();
+            $entityB = $revisionB->getRepository();
+            if ($entityA !== $entityB) {
+                return $recentTimestampsPerEntity->get($entityB->getId()) - $recentTimestampsPerEntity->get($entityA->getId());
+            } else {
+                return $revisionB->getId() - $revisionA->getId();
+            }
+        });
+        $collection = new ArrayCollection(iterator_to_array($iterator));
+        return $collection;
+    }
+
+    protected function isInSubject(EntityInterface $entity, TaxonomyTermInterface $term)
+    {
+        if (!$entity->getTaxonomyTerms()->isEmpty()) {
+            foreach ($entity->getTaxonomyTerms() as $tempTerm) {
+                if ($tempTerm->knowsAncestor($term)) {
+                    return true;
+                }
+            }
+        } else {
+            foreach ($entity->getParents('link') as $parent) {
+                if ($this->isInSubject($parent, $term)) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
+    protected function getEntities(TaxonomyTermInterface $term)
+    {
+        return $term->getAssociatedRecursive('entities');
     }
 
     protected function iterEntities(Collection $entities, Collection $collection, $callback)